+++ CHANGELOG.md
@@ -9,7 +9,6 @@
 - [](https://github.com/vegaprotocol/vega/pull/) - 
 - 
 ### 🛠 Improvements
-- [4413](https://github.com/vegaprotocol/vega/pull/4413) - Deploy to Devnet with Jenkins and remove drone
 - [4406](https://github.com/vegaprotocol/data-node/pull/4406) - Add changelog and project board Github actions and update linked PR action version
 - [4328](https://github.com/vegaprotocol/vega/pull/4328) - Unwrap the timestamps in reward payout event
 - [4330](https://github.com/vegaprotocol/vega/pull/4330) - Remove badger related code from the codebase
@@ -35,6 +34,7 @@
 - [4394](https://github.com/vegaprotocol/vega/pull/4394) - Add step to clear all events in integration tests
 - [4403](https://github.com/vegaprotocol/vega/pull/4403) - Fully remove expiry from withdrawals #4403
 - [4396](https://github.com/vegaprotocol/vega/pull/4396) - Add free form governance proposals
+- [4413](https://github.com/vegaprotocol/vega/pull/4413) - Deploy to Devnet with Jenkins and remove drone
 
 ### 🐛 Fixes
 - [4325](https://github.com/vegaprotocol/vega/pull/4325) - Remove state from the witness snapshot and infer it from votes