+++ CHANGELOG.md
@@ -7,7 +7,17 @@
 
 ### 🗑️ Deprecation
 - [](https://github.com/vegaprotocol/vega/pull/) -
--
+
+### 🛠 Improvements
+- [](https://github.com/vegaprotocol/vega/pull/) -
+
+### 🐛 Fixes
+- [](https://github.com/vegaprotocol/vega/pull/) -
+
+
+## 0.46.0
+*2021-11-22*
+
 ### 🛠 Improvements
 - [4431](https://github.com/vegaprotocol/vega/pull/4431) - Update Vega wallet to version 0.10.0
 - [4406](https://github.com/vegaprotocol/vega/pull/4406) - Add changelog and project board Github actions and update linked PR action version
@@ -47,6 +57,7 @@
 - [4395](https://github.com/vegaprotocol/vega/pull/4395) - Fix drone pipeline
 - [4398](https://github.com/vegaprotocol/vega/pull/4398) - Fix to set proper status on withdrawal errors
 - [4421](https://github.com/vegaprotocol/vega/issues/4421) - Fix to missing pending rewards in LNL checkpoint
+
 
 ## 0.45.5
 *2021-11-16*